requirements.txt
streamlit==1.35.0
python-docx==1.1.0
pypdf==4.0.1
openpyxl==3.1.2
xlrd==2.0.1
pandas==2.2.2
jieba==0.42.1
python-dotenv==1.0.1
reportlab==4.2.0
Pillow==10.4.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
